ADAS/requirements.txt
# Основные
torch
torchvision
ultralytics

# Для данных
roboflow
opencv-python
pillow
numpy

# Для визуализации
matplotlib
seaborn
pandas

# Утилиты
tqdm
pyyaml